Company Registration and w-full fix
The Company Registration Card is finished so now all that's left is making an Auth0 account and inputting everything into a Database.

Also edited the main.css file and added a #root so now the screen will default to full screen with w-full.

Will do the same for the reviewer once everything is finished.